+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Select original-language auto-generated subtitles when dubbed audio also has auto-generated subtitle tracks.
+
 
 ## [2.24.6] - 2026-09-04
 ### Fixed